fixed model tests?

+++ tests/models/Reader.test.js
@@ -1,11 +1,12 @@
 const tap = require('tap')
 const { destroyDB } = require('../integration/utils')
-const app = require('../../server').app
 const { Reader } = require('../../models/Reader')
 const parseurl = require('url').parse
 
-const test = async () => {
-  await app.initialize()
+const test = async app => {
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.initialize()
+  }
 
   const reader = Object.assign(new Reader(), {
     id: '123456789abcdef',
@@ -75,8 +76,10 @@ const test = async () => {
     await tap.equal(refObject.userId, undefined)
   })
 
-  await app.terminate()
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.terminate()
+  }
   await destroyDB(app)
 }
 
-test()
+module.exports = test
